revert fixes for diskio and ethtool
